--- IVT20241127-QCELLS_Q.PEAK-DUO-L-G8.3BFG-415-20241127_120616_824220393395101857_REAR_01.jsonld
@@ -1,27 +1,0 @@
-{
-  "@context": {
-    "mds": "https://cwrusdle.bitbucket.io/mds/"
-  },
-  "@id": "mds:29167",
-  "@type": "mds:CurrentVoltageResults",
-  "mds:SerialNumber": "824220393395101857",
-  "mds:CurrentVoltageSample": "p-type",
-  "mds:CurrentVoltageResultsTimestamp": "07:31.7",
-  "mds:ShortCircuitCurrent": 6.30234,
-  "mds:OpenCircuitVoltage": 46.81877,
-  "mds:CurrentAtMaxPower": 4.514477,
-  "mds:VoltageAtMaxPower": 40.887062,
-  "mds:MaximumPowerPoint": 184.583713,
-  "mds:FillFactor": 62.556366,
-  "mds:CellEfficiency": 8.615745,
-  "mds:ShuntResistance": 5091.239778,
-  "mds:SeriesResistance": 0.702843,
-  "mds:SunIntensity": 1.0,
-  "mds:SampleCellArea": 119.8,
-  "mds:SampleTotalArea": 21424.0,
-  "mds:NumberOfCellsPerString": 72,
-  "mds:NumberStrings": 2,
-  "mds:SampleResistivity": 0.954276,
-  "mds:CellReferenceConstant": 0.244789,
-  "mds:VoltageTemperatureCoefficient": 1.8221
-}
