--- IVT20241127-QCELLS_Q.PEAK-DUO-L-G8.3BFG-415-20241127_142458_824220393395101857_FRONT_01.jsonld
@@ -1,27 +1,0 @@
-{
-  "@context": {
-    "mds": "https://cwrusdle.bitbucket.io/mds/"
-  },
-  "@id": "mds:29169",
-  "@type": "mds:CurrentVoltageResults",
-  "mds:SerialNumber": "824220393395101857",
-  "mds:CurrentVoltageSample": "p-type",
-  "mds:CurrentVoltageResultsTimestamp": "26:15.1",
-  "mds:ShortCircuitCurrent": 10.2038,
-  "mds:OpenCircuitVoltage": 47.629947,
-  "mds:CurrentAtMaxPower": 9.496178,
-  "mds:VoltageAtMaxPower": 39.762754,
-  "mds:MaximumPowerPoint": 377.594178,
-  "mds:FillFactor": 77.693246,
-  "mds:CellEfficiency": 17.624822,
-  "mds:ShuntResistance": 5050.300157,
-  "mds:SeriesResistance": 0.208374,
-  "mds:SunIntensity": 1.0,
-  "mds:SampleCellArea": 119.8,
-  "mds:SampleTotalArea": 21424.0,
-  "mds:NumberOfCellsPerString": 72,
-  "mds:NumberStrings": 2,
-  "mds:SampleResistivity": 0.954276,
-  "mds:CellReferenceConstant": 0.244789,
-  "mds:VoltageTemperatureCoefficient": 1.8221
-}
