--- IVT20241209-QCELLS_Q.PEAK-DUO-L-G8.3BFG-415-20241209_160902_824220393395101857_CORED-REAR_02.jsonld
@@ -1,27 +1,0 @@
-{
-  "@context": {
-    "mds": "https://cwrusdle.bitbucket.io/mds/"
-  },
-  "@id": "mds:29173",
-  "@type": "mds:CurrentVoltageResults",
-  "mds:SerialNumber": "824220393395101857",
-  "mds:CurrentVoltageSample": "p-type",
-  "mds:CurrentVoltageResultsTimestamp": "10:17.9",
-  "mds:ShortCircuitCurrent": 6.672606,
-  "mds:OpenCircuitVoltage": 47.053141,
-  "mds:CurrentAtMaxPower": 4.858388,
-  "mds:VoltageAtMaxPower": 40.959385,
-  "mds:MaximumPowerPoint": 198.996607,
-  "mds:FillFactor": 63.381365,
-  "mds:CellEfficiency": 9.28849,
-  "mds:ShuntResistance": 5355.009716,
-  "mds:SeriesResistance": 0.678567,
-  "mds:SunIntensity": 1.0,
-  "mds:SampleCellArea": 119.8,
-  "mds:SampleTotalArea": 21424.0,
-  "mds:NumberOfCellsPerString": 72,
-  "mds:NumberStrings": 2,
-  "mds:SampleResistivity": 0.954276,
-  "mds:CellReferenceConstant": 0.2449,
-  "mds:VoltageTemperatureCoefficient": 1.8221
-}
