+++ data/IVT20241127-QCELLS_Q.PEAK-DUO-L-G8.3BFG-415-20241127_093908_824220393395101857_REAR_02.jsonld
@@ -1,0 +1,27 @@
+{
+  "@context": {
+    "mds": "https://cwrusdle.bitbucket.io/mds/"
+  },
+  "@id": "mds:29164",
+  "@type": "mds:CurrentVoltageResults",
+  "mds:SerialNumber": "824220393395101857",
+  "mds:CurrentVoltageSample": "p-type",
+  "mds:CurrentVoltageResultsTimestamp": "40:23.1",
+  "mds:ShortCircuitCurrent": 6.263255,
+  "mds:OpenCircuitVoltage": 46.782881,
+  "mds:CurrentAtMaxPower": 4.482143,
+  "mds:VoltageAtMaxPower": 40.91567,
+  "mds:MaximumPowerPoint": 183.389874,
+  "mds:FillFactor": 62.587602,
+  "mds:CellEfficiency": 8.56002,
+  "mds:ShuntResistance": 4914.296851,
+  "mds:SeriesResistance": 0.694166,
+  "mds:SunIntensity": 1.0,
+  "mds:SampleCellArea": 119.8,
+  "mds:SampleTotalArea": 21424.0,
+  "mds:NumberOfCellsPerString": 72,
+  "mds:NumberStrings": 2,
+  "mds:SampleResistivity": 0.954276,
+  "mds:CellReferenceConstant": 0.244789,
+  "mds:VoltageTemperatureCoefficient": 1.8221
+}
