+++ data/IVT20241209-QCELLS_Q.PEAK-DUO-L-G8.3BFG-415-20241209_162131_824220393395101857_CORED-FRONT_01.jsonld
@@ -1,0 +1,27 @@
+{
+  "@context": {
+    "mds": "https://cwrusdle.bitbucket.io/mds/"
+  },
+  "@id": "mds:29174",
+  "@type": "mds:CurrentVoltageResults",
+  "mds:SerialNumber": "824220393395101857",
+  "mds:CurrentVoltageSample": "p-type",
+  "mds:CurrentVoltageResultsTimestamp": "22:47.0",
+  "mds:ShortCircuitCurrent": 10.274601,
+  "mds:OpenCircuitVoltage": 47.843066,
+  "mds:CurrentAtMaxPower": 9.643962,
+  "mds:VoltageAtMaxPower": 39.785474,
+  "mds:MaximumPowerPoint": 383.689616,
+  "mds:FillFactor": 78.054164,
+  "mds:CellEfficiency": 17.909336,
+  "mds:ShuntResistance": 5683.250997,
+  "mds:SeriesResistance": 0.225377,
+  "mds:SunIntensity": 1.0,
+  "mds:SampleCellArea": 119.8,
+  "mds:SampleTotalArea": 21424.0,
+  "mds:NumberOfCellsPerString": 72,
+  "mds:NumberStrings": 2,
+  "mds:SampleResistivity": 0.954276,
+  "mds:CellReferenceConstant": 0.2449,
+  "mds:VoltageTemperatureCoefficient": 1.8221
+}
