--- data_flow.d2
@@ -1,75 +1,0 @@
-# Data Sources
-group pvmcf_data_grid: {
-  label: "Photovoltaic Module Charecterization Facility Data Sources"
-  module_metadata
-  module_status
-  measurement_settings
-  sinton_iv_metadata
-  sinton_iv_results
-  spire_iv_metadata
-  spire_iv_results
-  dark_iv_metadata
-  el_metadata
-  ir_indoor_metadata
-  uvf_indoor_metadata
-  ir_outdoor_metadata
-  scanner_jpg_metadata
-  scanner_nc_metadata
-  grid-rows: 3
-}
-
-group timeseries_data_grid: {
-  label: "Time Series and System Data Sources"
-  8152-railpads
-  8157
-  evergreen
-  fgcu
-  forum-the-wine-company
-  ilte
-  met-stations
-  nist-pv
-  pv-lifetime
-  slte
-  sunsmart
-  vcad
-  FSEC HISTORICAL DATA
-  grid-rows: 3
-}
-
-group literature_data_grid: {
-  label: "Historical and Literature Data Sources"
-  zotera
-  google-scholar
-  pv-education
-  research-gate
-  science-open
-  science-direct
-  grid-rows: 3
-}
-
-# Components (Databases and Services)
-fsecdatabase: {
-  label: "FSEC Database"
-  shape: cylinder
-}
-pvmfc_database: {
-  label: "PVMCF Database"
-  shape: cylinder
-}
-
-time_series_database: {
-  label: "Systems Database"
-  shape: cylinder
-}
-
-literature_database: {
-  label: "Literature Database"
-  shape: cylinder
-}
-
-# Connections and Data Flows
-pvmfc_database -> fsecdatabase: "Charecterization and Exposur Testing"
-time_series_database -> fsecdatabase: "Time Series Data"
-literature_database -> fsecdatabase: "Historical and Literature Data"
-group pvmcf_data_grid -> pvmfc_database: "Instrument Data Parser"
-group timeseries_data_grid -> time_series_database: "Argonaut"
